requirements.txt
litellm
google-adk
openai
python-docx
PyPDF2
python-dotenv